+++ src/AST/Nodes/ClassDefinition.h
@@ -39,7 +39,7 @@ class ClassDefinition : public ASTNode {
 			m_PARENTCLASSNAME = std::nullopt;
 		}
 
-		std::ostream& prettyPrint(std::ostream& os, int indentation_level = 0) const override {
+		std::ostream& prettyPrint(std::ostream& os, size_t indentation_level = 0) const override {
 			std::string indent(indentation_level * PRETTYPRINT_INDENTATION_AMOUNT, ' ');
 			os << indent << "(ClassDefinition\n"
 				<< indent << "  @class " << m_CLASSNAME;
