+++ src/AST/Nodes/ParameterExpansion.h
@@ -26,7 +26,7 @@ class ParameterExpansion : public StringType {
 			m_EXPANSIONBEGIN = expansionBegin;
 		}
 
-		std::ostream& prettyPrint(std::ostream& os, int indentation_level = 0) const override {
+		std::ostream& prettyPrint(std::ostream& os, size_t indentation_level = 0) const override {
 			std::string indent(indentation_level * PRETTYPRINT_INDENTATION_AMOUNT, ' ');
 			os << indent << "(ParameterExpansion\n"
 				<< indent << "  " << m_EXPANSIONBEGIN;
